requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
ConfigArgParse==1.2.3
Flask==1.1.2
gevent==20.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
locustio==0.14.6
MarkupSafe==1.1.1
msgpack==1.0.0
psutil==5.7.0
pymongo==3.10.1
pyzmq==19.0.0
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
